refactor: Update to the data flow figure and remove the solidifier workerpool
